Saving invoker name in CConsole::getConsoleInstance
In order to implement ability in CConsole to filter based on invoker module, I've added a string param to CConsole::getConsoleInstance() which is saved as loggerName. So CConsole always knows WHICH module invoked it for logging. Modules should use CConsole::getConsoleInstance() from now on, with their module name, to use the per-module filtering ability.
Filtering is not implemented yet, it will be done in next commit.